Remove sidebar maxWidth
The maxWidth was applying only to the List it contained, so the sidebar wasn't prevented form being resized larger than maxWidth. When this happens, the List appears at it's maxWidth centred in the sidebar.

I did try setting it on the sidebar directly (when it's added to the NavigationView), but this had the same effect (the NavigationView could still be resized larger than maxWidth, with the same effect as before).